Address clippy warnings; prefer &Path and minor fixes
Replace &PathBuf parameters with &Path and convert clones to to_path_buf where a PathBuf is required; initialize WINDOWPLACEMENT with an explicit length field; remove unnecessary casts for BMP pixel byte calculations; use is_multiple_of for checkerboard parity checks; pass vf as a &str in ffmpeg args; load tray icon resource via without_provenance instead of a pointer cast; update CHANGELOG to note the clippy fixes.

+++ src/office_preview.rs
@@ -195,7 +195,7 @@ mod tests {
     /// one: a `BITMAPINFO` and its pixels with a file header in front of them. The
     /// alpha byte is zero, which is what Excel writes there.
     fn bmp_bytes(width: u32, height: u32, color: [u8; 4]) -> Vec<u8> {
-        let pixel_bytes = (width * height * 4) as u32;
+        let pixel_bytes = width * height * 4;
         let mut dib = Vec::new();
         dib.extend_from_slice(&40u32.to_le_bytes()); // header size
         dib.extend_from_slice(&(width as i32).to_le_bytes());
